Rename module and update imports
The module `afdko_tools.py` was renamed to `cffsubr.py` to more accurately reflect its content and use. The import statement in `font.py` was accordingly updated to match the new name.
